test: move combo-box item template tests to ITs
